Simplify `OLAPStore` interface
