+++ runtime/drivers/clickhouse/model_executor_localfile_self.go
@@ -17,7 +17,7 @@ import (
 
 type localFileToSelfExecutor struct {
 	fileStore drivers.Handle
-	c         *connection
+	c         *Connection
 }
 
 var _ drivers.ModelExecutor = &selfToSelfExecutor{}
@@ -135,7 +135,7 @@ func (e *localFileToSelfExecutor) Execute(ctx context.Context, opts *drivers.Mod
 		}
 	} else if stagingTableName != tableName {
 		// Rename the staging table to the final table name
-		err = olapForceRenameTable(ctx, e.c, stagingTableName, false, tableName)
+		err = e.c.olapForceRenameTable(ctx, stagingTableName, false, tableName)
 		if err != nil {
 			return nil, fmt.Errorf("failed to rename staged model: %w", err)
 		}
